--- resources/views/livewire/admin/create-user.blade.php
@@ -1,3 +1,0 @@
-<div>
-    {{-- In work, do what you enjoy. --}}
-</div>
